Add files via upload
This is my capstone project to learn python for data analysis and data visualisation. In this project i will be taking a look at 2008 financial market crash. please keep in mind that this is not any sort of financial expertise, this was just done to learn python for data science.